git-p4: improve branch option handling
It is sometimes useful (much quicker) to request commands only operate
on a single branch.

The P4Sync command has been updated to handle self.branch being None for
consitency with the P4Submit.

The P4Rebase command has been given a branch option which is forwarded
to the P4Sync command it runs.

The P4Submit command has been simplified to not call P4Sync itself, it
lets P4Rebase do it instead (now that the branch can be handled).  This
fixes an issue where P4Submit does a sync of the requested branch and
then does a rebase which does a sync of all branches.

Signed-off-by: Andrew Oakley <aoakley@roku.com>