refs/files-backend: convert many internals to struct object_id
Convert many of the internals of the files backend to use struct
object_id.  Avoid converting public APIs to limit the scope of the
changes.

Convert one use of get_sha1_hex to parse_oid_hex, and rely on the fact
that a strbuf will be NUL-terminated and that parse_oid_hex will fail on
truncated input to avoid the need to check for an explicit length.

This is a requirement to convert parse_object later on.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>