remote: warn loud and clear when .git/branches/ is *still* used
Since at least 5751f49010e (Move remote parsing into a library file out
of builtin-push., 2007-05-12), we strongly prefer remotes and upstream
branches to be specified in the config rather than .git/remotes/ and
.git/branches/.

For some time, we still retained compatibility with Cogito (which was an
excellent project demonstrating many useful features that eventually
were integrated into Git proper) by creating the branches/ subdirectory
upon git-init.

In preparation for removing support for this ancient compatibility that
should no longer be necessary, let's warn everybody who is still using
that feature (if there is anybody left, that is).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>