docs: fix literal quoted spaces
When compiling the documentation, asciidoc thinks a backtick surrounded
by whitespace shouldn't be interpreted as marking the start or end of a
literal.  In most cases, that's useful behaviour, but in the git-pull
documentation the space is clearly intended to be part of the monospace
formatted text.

Instead, use + to avoid asciidoc's literal passthrough, and encode the
space as {sp}.  In particular, this means asciidoc will correctly detect
the end of the monospace formatting, rather than having it continue past
the backtick.

Signed-off-by: Adam Dinwoodie <adam@dinwoodie.org>